Refactored codebase to be more modular
