+++ .python-version
@@ -1,1 +1,1 @@
-3.10
+3.11.5
