+++ pdf_data_viewer/core/__init__.py
@@ -1,0 +1,1 @@
+"""Core functionality modules for the PDF Data Viewer application."""
