+++ pdf_data_viewer/database/__init__.py
@@ -1,0 +1,1 @@
+"""Database modules for the PDF Data Viewer application."""
