+++ pdf_data_viewer/ui/__init__.py
@@ -1,0 +1,1 @@
+"""UI modules for the PDF Data Viewer application."""
