+++ pdf_data_viewer/utils/__init__.py
@@ -1,0 +1,1 @@
+"""Utility modules for the PDF Data Viewer application."""
